refactor : Application class 최상위 패키지로 이동. Controller class 구현
