+++ src/main/java/racinggame/Application.java
@@ -1,0 +1,10 @@
+package racinggame;
+
+import racinggame.controller.RacingGameController;
+
+public class Application {
+    public static void main(String[] args) {
+        RacingGameController racingGameController = new RacingGameController();
+        racingGameController.run();
+    }
+}
